+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="v3.70.2"></a>
+## [v3.70.2] - 2025-09-25
+### Bug Fixes
+- **newrelic_one_dashboard:** add plan-time validation for `account_ids` in `variable{}` block ([#2930](https://github.com/newrelic/terraform-provider-newrelic/issues/2930))
+
+<a name="v3.70.1"></a>
+## [v3.70.1] - 2025-09-25
+### Bug Fixes
+- **cloud:** make subnet optional in oci module ([#2938](https://github.com/newrelic/terraform-provider-newrelic/issues/2938))
+
+<a name="v3.70.0"></a>
+## [v3.70.0] - 2025-09-24
+### Features
+- **cloud:** add `newrelic_cloud_oci_link account` resource, associated onboarding modules ([#2936](https://github.com/newrelic/terraform-provider-newrelic/issues/2936))
+
 <a name="v3.69.1"></a>
 ## [v3.69.1] - 2025-09-23
 ### Documentation Updates
@@ -2442,7 +2457,10 @@ new synthetics resources use GraphQL API schema
 ## [v0.1.1] - 2017-08-02
 <a name="v0.1.0"></a>
 ## v0.1.0 - 2017-06-21
-[Unreleased]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.69.1...HEAD
+[Unreleased]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.70.2...HEAD
+[v3.70.2]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.70.1...v3.70.2
+[v3.70.1]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.70.0...v3.70.1
+[v3.70.0]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.69.1...v3.70.0
 [v3.69.1]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.69.0...v3.69.1
 [v3.69.0]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.68.0...v3.69.0
 [v3.68.0]: https://github.com/newrelic/terraform-provider-newrelic/compare/v3.67.0...v3.68.0